+++ src/examples/tx_thread.nim
@@ -1,3 +1,6 @@
+when not compileOption("threads"):
+  {.fatal: "Must be compiled with --threads:on".}
+
 import std/[random, strformat, strutils, times]
 import yottadb
 import ydbutils
@@ -6,52 +9,48 @@ import malebolgia
 # Demonstrates that the txid counter ^CNT("M") stays in a consitent state
 # All txid^s must be consecutive.
 
-when not compileOption("threads"):
-  {.fatal: "Must be compiled with --threads:on".}
 
 const
   NUM_OF_THREADS = 4
-  ITERATIONS = 10
+  ITERATIONS = 25
   THS = "M"
   GLOBAL = "^TX" & THS
 
-# Calculate the highest value to enforce a timeout
-let minFibonacci = calcFibonacciValueFor1000ms(50)
-let maxFibonacci = calcFibonacciValueFor1000ms(1200)
+kill: @GLOBAL
+kill: ^CNT
 
-
-# tptoken supplied from yottadb. Must be propagated with each call to ydb_set,....
-proc myTxnMT*(tptoken: uint64; buff: ptr ydb_buffer_t; param: pointer): cint  {.cdecl.} =
-  let tn = $cast[cstring](param)
-  let restarted = parseInt(ydb_get("$TRESTART", tptoken=tptoken)) # How many times the proc was called from yottadb
-
-  # Do cpu intense work
-  let (ms, fibresult) = timed_rc:
-    let fib = rand(minFibonacci .. maxFibonacci)
-    fibonacci_recursive(fib) # do some cpu intense work
-
-  try:  
-    # Increment a serial number and save application data
-    let txid = ydb_increment("^CNT", @[THS, tn], 1, tptoken)
-    let data = fmt"restarts:{restarted}, fib:{fib} result:{fibresult} time:{ms}, tptoken:{tptoken}"
-    ydb_set(GLOBAL, @[$txid, tn], $data, tptoken)
-  except:
-    # Retry a aborted transaction one time, otherwise roll back
-    if restarted == 0: return YDB_TP_RESTART else: return YDB_TP_ROLLBACK
-
-  return YDB_OK # commit the transaction
+let minFibonacci = calcFibonacciValueFor1000ms(10)
+let maxFibonacci = calcFibonacciValueFor1000ms(300)
 
 proc worker(tn: int, iterations: int) =
-  ydb_set("$ZMAXTPTIME", value="1") # Set transaction timeout to 1 second
+  setvar: $ZMAXTPTIME = 1 # Set transaction timeout to 1 second
+
   for cnt in 0..<ITERATIONS:
-    let (ms, rc) = timed_rc:
-      ydb_tp_mt(myTxnMT, $tn, $cnt)
+    # Do cpu intense work
+    let (ms, fibresult) = timed_rc:
+      let fib = rand(minFibonacci .. maxFibonacci)
+      fibonacci_recursive(fib) # do some cpu intense work
+
+    # Save data
+    let info = $tn & "," & $fibresult & "," & $ms
+    let rc = TransactionMT(info):
+      let info = $cast[cstring](param)
+      let tnr = info.split(",")[0]
+      let fibresult = info.split(",")[1]
+      let ms = info.split(",")[2]
+      let restarted = parseInt(ydb_get("$TRESTART", tptoken=tptoken)) # How many times the proc was called from yottadb
+      let txid = ydb_increment("^CNT", @[THS, tnr], 1, tptoken)
+      let data = fmt"tn:{tnr}, restarts:{restarted}, fibresult:{fibresult}, ms:{ms}, tptoken:{tptoken}"
+      ydb_set(GLOBAL, @[$txid, $tnr], $data, tptoken)
     
     if rc == YDB_OK:
-      let txid = ydb_get("^CNT", @[THS, $tn]) # get last transaction id for this thread
-      var data = newYdbVar(GLOBAL, @[$txid, $tn])
-      data[] = data.value & " overall-time:" & $ms # append overall
-      echo "tn:", tn, " cnt:", cnt, " txid:", txid, " data: ", data
+      let rc = TransactionMT1($tn & "," & $ms):
+        let info = $cast[cstring](param)
+        let tn = info.split(",")[0]
+        let ms = info.split(",")[1]
+        let txid = ydb_get("^CNT", @[THS, $tn], tptoken=tptoken) # get last transaction id for this thread
+        var data = newYdbVar(GLOBAL, @[$txid, $tn], tptoken=tptoken)
+        data[] = data.value & " overall-time:" & $ms # append overall
     else:
       # Should not happen
       echo "Error ", rc, " after running ydb_tp_mt"
@@ -64,4 +63,13 @@ when isMainModule:
     for tn in 0..<NUM_OF_THREADS:
       m.spawn worker(tn, ITERATIONS)
 
-  echo "all finished"
+  echo "All threads finished"
+
+  listVar("^TXM")
+  # test results
+  for i in 0..<NUM_OF_THREADS:
+    let txs = getvar ^CNT("M", i).int
+    assert txs == ITERATIONS
+  for i in 1..ITERATIONS:
+    for tn in 0..<NUM_OF_THREADS:
+      assert 1 == data ^TXM(i, tn)